Played around and improved chessBoard mini project. It is now up to date, very customizable, and very well commented
Moved pomodoro.py and changed its name to `complexPomodoro.py`

Finally, and most importantly, worked and tested `PEMDASsimplifier.py`, module which solves the equation step by step. Although it is not extensively
tested, and instances like 2(2) produce an incorrect result, and the procedure is not outputted, it is a very big step towards the LaTEX equation
procedure generator.